+++ BlueControls/BlueControls/Interfaces/IMouseAndKeyHandle.cs
@@ -21,19 +21,19 @@ namespace BlueControls.Interfaces {
 
         #region Methods
 
-        bool KeyUp(object sender, System.Windows.Forms.KeyEventArgs e, double zoom, double shiftX, double shiftY);
+        bool KeyUp(object sender, System.Windows.Forms.KeyEventArgs e, float zoom, float shiftX, float shiftY);
 
-        bool MouseDown(object sender, System.Windows.Forms.MouseEventArgs e, double zoom, double shiftX, double shiftY);
+        bool MouseDown(object sender, System.Windows.Forms.MouseEventArgs e, float zoom, float shiftX, float shiftY);
 
-        bool MouseMove(object sender, System.Windows.Forms.MouseEventArgs e, double zoom, double shiftX, double shiftY);
+        bool MouseMove(object sender, System.Windows.Forms.MouseEventArgs e, float zoom, float shiftX, float shiftY);
 
-        bool MouseUp(object sender, System.Windows.Forms.MouseEventArgs e, double zoom, double shiftX, double shiftY);
+        bool MouseUp(object sender, System.Windows.Forms.MouseEventArgs e, float zoom, float shiftX, float shiftY);
 
         #endregion
 
         //Function BlueCreativePad_MouseWheel(sender As Object, e As System.Windows.Forms.MouseEventArgs) As Boolean
         //Function MouseClick(sender As Object, e As System.Windows.Forms.MouseEventArgs) As Boolean
-        //Function MouseDoubleClick(sender As Object, e As System.Windows.Forms.MouseEventArgs) As Boolean
+        //Function MousefloatClick(sender As Object, e As System.Windows.Forms.MouseEventArgs) As Boolean
         //Function MouseEnter(sender As Object, e As System.EventArgs) As Boolean
         //Function MouseHover(sender As Object, e As System.EventArgs) As Boolean
         //Function MouseLeave(sender As Object, e As System.EventArgs) As Boolean
